requirements.test.txt
-r requirements.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-vcr==1.0.2
cryptography==41.0.7
